change: ファイルを各フォルダに再配置
