+++ src/components/PokemonThumbnails.jsx
